FEM: mesh package, move all mesh modules in new mesh package
