+++ src/Mod/Fem/FemToolsCcx.py
@@ -803,7 +803,7 @@ def get_refshape_type(fem_doc_object):
     # in GUI defined material_obj could have no RefShape and RefShapes could be different type
     # we're going to need the RefShapes to be the same type inside one fem_doc_object
     # TODO: check if all RefShapes inside the object really have the same type
-    import FemMeshTools
+    import femmesh.meshtools as FemMeshTools
     if hasattr(fem_doc_object, 'References') and fem_doc_object.References:
         first_ref_obj = fem_doc_object.References[0]
         first_ref_shape = FemMeshTools.get_element(first_ref_obj[0], first_ref_obj[1][0])
