+++ src/Mod/Fem/femsolver/z88/writer.py
@@ -28,7 +28,7 @@ __url__ = "http://www.freecadweb.org"
 #  @{
 
 import FreeCAD
-import FemMeshTools
+import femmesh.meshtools as FemMeshTools
 import feminterface.importZ88Mesh as importZ88Mesh
 import FemInputWriter
